// Helpers/Types.tsx
export type createUserDefaultValue = {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  password: string;
  country: string;
  terms: boolean;
};
// ...
export type Tseller = {
  location: string;
  accId: string | null;
  paypal: string | null;
  owner: string;
};
export type TStoreId = {
  name: string;
  _id: string;
  logo: string;
  sales: number;
  owner: Tseller;
  location: string;
  currency: string;
  disabled: boolean;
  domesticShipping: {
    standard: number;
    express: number;
  };
};
export type TStoreId1 = {
  name: string;
  _id: string;
  logo: string;
  balance: number;
  owner: Tseller;
  location: string;
  currency: string;
  disabled: boolean;
};
export type TRefunds = {
  _id: string;
  storeId: TStoreId;
  userId: createUserDefaultValue;
  status: string;
  reason: string;
};
export type TTransaction = {
  _id: string;
  account: string;
  method: string;
  amount: number;
  status: string;
  createdAt: Date;
}
export type TSellerStore = {
  _id: string;
  accId: string;
  paypal: string;
  name: string;
  isActive: boolean;
  owner: createUserDefaultValue;
  documentType: string;
  location: string;
  documentId: string;
  storeId: TStoreId1;
  file: string;
  isVerified: boolean;
  isPausedPayout: boolean;
  account: string;
  websiteUrl: string;
};